+++ src/main/java/de/thomas_oster/liblasercut/FloatPowerSpeedFocusProperty.java
@@ -38,7 +38,6 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   /**
    * Sets the Laserpower. Valid values are from 0 to 100.
    * In 3d-Raster mode, the intensity is scaled to this power setting
-   * @param power 
    */
   public void setPower(float power)
   {
@@ -54,7 +53,6 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
 
   /**
    * Sets the speed for the Laser. Valid values is from 0 to 100
-   * @param speed 
    */
   public void setSpeed(float speed)
   {
@@ -113,15 +111,15 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) this.getPower();
+      return this.getPower();
     }
     else if ("speed".equals(name))
     {
-      return (Float) this.getSpeed();
+      return this.getSpeed();
     }
     else if ("focus".equals(name))
     {
-      return (Float) this.getFocus();
+      return this.getFocus();
     }
     return null;
   }
@@ -152,11 +150,11 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
   if ("power".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 0f;
+      return 0f;
     }
     else if ("focus".equals(name))
     {
@@ -177,11 +175,11 @@ public class FloatPowerSpeedFocusProperty implements LaserProperty
   {
     if ("power".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("speed".equals(name))
     {
-      return (Float) 100f;
+      return 100f;
     }
     else if ("focus".equals(name))
     {
